Make calculator work using handleCalculation(arg)
handleCalculation(arg) gets called from handleEvent(e) irrespective
of the type of the event (whether it be pointer or keyboard).

From the handleCalculation(arg) function, a switch statement makes the
calculator work and allows for the following operations:
- The calculate does not evaluate more than a single pair of numbers
at a time (e.g., 12 + 7 - leading to an answer of 19 -), therefore,
it does not adhere to the operator precedence.
- Answers with long decimals are rounded off so that they don't overflow
the display (some use a fixed set of significant figures and other long
ones may be scientific notation).
- Pressing "equals" or "any operator" before a calcualtion gives an
error and the error has been customized in a fun way so that random
words are thrown at the user for committing errors.